% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffpo.R
\name{add_grid}
\alias{add_grid}
\title{Grid adder for dataframes}
\usage{
add_grid(df, grid)
}
\arguments{
\item{df}{\code{data.frame} object to which the grid will be added.}

\item{grid}{Grid vector.}
}
\value{
\code{data.frame} with the grid added.
}
\description{
It prepared the partially observed data to be inputed in the \code{ffpo} function.
This function should only be used when the \code{bidimensional_grid}
parameter of the \code{ffpo} function is \code{FALSE}.
}
\seealso{
\code{\link{ffpo}}
}
