% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B2XZG.R
\name{B2XZG}
\alias{B2XZG}
\title{Multidimensional B2XZG function}
\usage{
B2XZG(B, pord = c(2, 2), c = c(10, 10))
}
\arguments{
\item{B}{Design matrix.}

\item{pord}{Order of the matrix of differences. We are going to use a
bidimensional penalization.}

\item{c}{Polynomial order.}
}
\value{
.
}
\description{
Multidimensional B2XZG function
}
