% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B2XZG_1d.R
\name{B2XZG_1d}
\alias{B2XZG_1d}
\title{One dimensional B2XZG}
\usage{
B2XZG_1d(B, pord = c(2), c = c(10))
}
\arguments{
\item{B}{Design matrix.}

\item{pord}{Order of the difference matrix.}

\item{c}{Polynomial order.}
}
\value{
.
}
\description{
One dimensional B2XZG
}
