% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffpo_2d.R
\name{B2XZG_2d}
\alias{B2XZG_2d}
\title{B2XZG  2-dimensional}
\usage{
B2XZG_2d(B, pord = c(2, 2), c = c(10, 10))
}
\arguments{
\item{B}{.}

\item{pord}{.}

\item{c}{.}
}
\value{
.
}
\description{
This is used as part of the ffpo_2d function
}
