% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspline.R
\name{bspline}
\alias{bspline}
\title{B-spline generator}
\usage{
bspline(X., XL., XR., NDX., BDEG.)
}
\arguments{
\item{X.}{Points where you are going to evaluate the function.}

\item{XL.}{Min x (- epsilon)}

\item{XR.}{Max x (+ epsilon)}

\item{NDX.}{Number of basis - BDEG.}

\item{BDEG.}{Degree of the polynomial.}
}
\value{
A \code{list} where the first element is the design matrix of the
B-spline and the second element is a list with the different knots.
}
\description{
B-spline generator
}
