% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data2B.R
\name{Data2B}
\alias{Data2B}
\title{Data to B matrix}
\usage{
Data2B(X, M, nbasis = c(30, 30, 30), bdeg = c(3, 3, 3), sub = 500, lim = NULL)
}
\arguments{
\item{X}{.}

\item{M}{.}

\item{nbasis}{.}

\item{bdeg}{.}

\item{sub}{.}

\item{lim}{.}
}
\value{
.
}
\description{
Data to B matrix
}
