% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data2B_simpson.R
\name{Data2B_simpson}
\alias{Data2B_simpson}
\title{Title}
\usage{
Data2B_simpson(
  X,
  M,
  nbasis = c(30, 30, 30),
  bdeg = c(3, 3, 3),
  sub = 25,
  lim = NULL
)
}
\arguments{
\item{X}{.}

\item{M}{.}

\item{nbasis}{.}

\item{bdeg}{.}

\item{sub}{.}

\item{lim}{.}
}
\value{
.
}
\description{
Title
}
