% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2B_simpson_classification.R
\name{data2B_simpson_classification}
\alias{data2B_simpson_classification}
\title{data2B simpson for classification problems}
\usage{
data2B_simpson_classification(X, M, grid, nbasis = c(80, 80), bdeg = c(3, 3))
}
\arguments{
\item{X}{.}

\item{M}{.}

\item{grid}{.}

\item{nbasis}{.}

\item{bdeg}{.}
}
\value{
.
}
\description{
data2B simpson for classification problems
}
