% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generator.R
\name{data_generator_vd}
\alias{data_generator_vd}
\title{Data generator function for the variable domain case.}
\usage{
data_generator_vd(
  N = 100,
  J = 100,
  nsims = 1,
  Rsq = 0.95,
  aligned = TRUE,
  multivariate = FALSE,
  beta_index = 1,
  use_x = FALSE,
  use_f = FALSE,
  seed = NULL
)
}
\arguments{
\item{N}{Number of subjects.}

\item{J}{Number of maximum observations per subject.}

\item{nsims}{Number of simulations per the simulation study.}

\item{Rsq}{Variance of the model.}

\item{aligned}{If the data that will be generated is aligned or not.}

\item{multivariate}{If TRUE, the data is generated with 2 functional variables.}

\item{beta_index}{Index for the beta.}

\item{use_x}{If the data is generated with x.}

\item{use_f}{If the data is generated with f.}

\item{seed}{Seed for reproducibility.}
}
\value{
A list containing the following components:
\itemize{
\item y: \code{vector} of length N containing the response variable.
\item X_s: \code{matrix} of non-noisy functional data for the first functional covariate.
\item X_se: \code{matrix} of noisy functional data for the first functional covariate
\item Y_s: \code{matrix} of non-noisy functional data for the second functional covariate (if multivariate).
\item Y_se: \code{matrix} of noisy functional data for the second covariate (if multivariate).
\item x1: \code{vector} of length N containing the non-functional covariate (if use_x is TRUE).
\item x2: \code{vector} of length N containing the observed values of the smooth term (if use_f is TRUE).
\item smooth_term: \code{vector} of length N containing a smooth term (if use_f is TRUE).
\item Beta: \code{array} containing the true functional coefficients.
}
Where N is the number of subjects and maxM is the maximum number of observations per subject.
}
\description{
This function is for internal use.
}
