% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{dg}
\alias{dg}
\title{Data generator function}
\usage{
dg(
  N = 100,
  J = 100,
  nsims = 1,
  Rsq = 0.95,
  aligned = TRUE,
  multivariate = FALSE,
  use_x = FALSE,
  use_f = FALSE
)
}
\arguments{
\item{N}{Number of subjects.}

\item{J}{Number of maximum observations per subject.}

\item{nsims}{Number of simulations per the simulation study.}

\item{Rsq}{.}

\item{aligned}{If the data is aligned or not.}

\item{multivariate}{If TRUE, the data is generated with 2 variables.}

\item{use_x}{If the data is generated with x.}

\item{use_f}{If the data is generated with f.}
}
\value{
Example data.
}
\description{
Data generator function
}
