% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffpo.R
\name{ffpo}
\alias{ffpo}
\title{ffpo}
\usage{
ffpo(X, grid, bidimensional_grid = FALSE, nbasis = c(30, 30), bdeg = c(3, 3))
}
\arguments{
\item{X}{.}

\item{grid}{.}

\item{bidimensional_grid}{.}

\item{nbasis}{.}

\item{bdeg}{.}
}
\value{
.
}
\description{
ffpo
}
