% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffvd.R
\name{ffvd}
\alias{ffvd}
\title{Defining variable domain functional data terms in vd_fit formulae}
\usage{
ffvd(X, grid, nbasis = c(30, 50, 30), bdeg = c(3, 3, 3))
}
\arguments{
\item{X}{variable domain functional covariate \code{matrix}.}

\item{grid}{observation points of the variable domain functional covariate.
If not provided, it will be \code{1:ncol(X)}.}

\item{nbasis}{number of basis to be used.}

\item{bdeg}{degree of the basis to be used.}
}
\value{
the function is interpreted in the formula of a \code{VDPO} model.
\code{list} containing the following elements:
\itemize{
\item An item named \code{B} design matrix.
\item An item named \code{X_hat} smoothed functional covariate.
\item An item named \code{L_Phi} and \code{B_T} 1-dimensional marginal B-spline basis used for the functional coefficient.
\item An item named \code{M} matrix object indicating the observed domain of the data.
\item An item named \code{nbasis} number of basis used.
}
}
\description{
Auxiliary function used to define \code{ffvd} terms within \code{vd_fit} model
formulae.
}
