% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generator.R
\name{generate_exp_surface}
\alias{generate_exp_surface}
\title{Generate exponential coefficient surface}
\usage{
generate_exp_surface(x, y)
}
\arguments{
\item{x}{Grid points for x-axis}

\item{y}{Grid points for y-axis}
}
\value{
Matrix containing the coefficient surface values
}
\description{
Generate exponential coefficient surface
}
