% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generator.R
\name{generate_saddle_surface}
\alias{generate_saddle_surface}
\title{Generate saddle-shaped coefficient surface}
\usage{
generate_saddle_surface(x, y)
}
\arguments{
\item{x}{Grid points for x-axis}

\item{y}{Grid points for y-axis}
}
\value{
Matrix containing the coefficient surface values
}
\description{
Generate saddle-shaped coefficient surface
}
