% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inprod.R
\name{inprod}
\alias{inprod}
\title{Inner product for fdobejct}
\usage{
inprod(
  fdobj1,
  fdobj2 = NULL,
  Lfdobj1 = fda::int2Lfd(0),
  Lfdobj2 = fda::int2Lfd(0),
  rng = range1,
  wtfd = 0
)
}
\arguments{
\item{fdobj1}{.}

\item{fdobj2}{.}

\item{Lfdobj1}{.}

\item{Lfdobj2}{.}

\item{rng}{.}

\item{wtfd}{.}
}
\value{
.
}
\description{
Inner product for fdobejct
}
