% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_inprod.R
\name{partial_inprod}
\alias{partial_inprod}
\title{Partial inner product}
\usage{
partial_inprod(n_intervals, knots1, knots2, bdeg, spline_domain, rng)
}
\arguments{
\item{n_intervals}{Number of intervals where we are going to integrate.
This should be an even number.}

\item{knots1}{First set of nodes.}

\item{knots2}{Seconds set of nodes.}

\item{bdeg}{Degree of the basis. This should be a two dimensional vector.}

\item{spline_domain}{Domain where the spline is defined.}

\item{rng}{Integration limits. This should be a two dimensional vector.}
}
\value{
Matrix with the value of the integral.
}
\description{
Partial inner product
}
