% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.po_2d_fit}
\alias{plot.po_2d_fit}
\title{Plot method for partially observed bidimensional functional regression fits}
\usage{
\method{plot}{po_2d_fit}(x, beta_index = 1, ...)
}
\arguments{
\item{x}{an object of class \code{po_2d_fit}, as returned by
\code{\link{po_2d_fit}}.}

\item{beta_index}{an integer selecting which coefficient surface to plot.
Default is 1.}

\item{...}{currently ignored, included for compatibility with the generic.}
}
\value{
A \code{ggplot2} object displaying the estimated coefficient surface.
}
\description{
Displays an estimated bidimensional functional coefficient of a
\code{po_2d_fit} object as a heatmap over its two-dimensional domain.
}
\seealso{
\code{\link{po_2d_fit}}
}
