% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_ci}
\alias{plot_ci}
\title{Plot Functional Curves with Confidence Intervals}
\usage{
plot_ci(object, beta_index = 1, curves)
}
\arguments{
\item{object}{An object of class \code{'vd_fit'} or similar, containing the fitted model results, Beta estimates, and evaluation details.}

\item{beta_index}{An integer specifying which Beta coefficient matrix to use. Default is 1.}

\item{curves}{A numeric vector specifying the indices of the curves (rows) to plot.}
}
\value{
A \code{ggplot2} object displaying the Beta estimates and confidence intervals for the specified curves.
}
\description{
Generates a plot of functional Beta estimates for specified curves, along with their 95\% confidence intervals.
This function computes the 95\% confidence intervals for each curve based on the covariance matrix and the fitted values
from the provided object. The resulting plot includes estimated curves, confidence interval ribbons, and a legend
distinguishing the curves.
}
\examples{
\donttest{
if (requireNamespace("ggplot2", quietly = TRUE)) {
  # set seed for reproducibility
  set.seed(42)

  # generate variable domain functional data and fit the model
  data <- data_generator_vd(N = 100, J = 100, beta_index = 1)
  res <- vd_fit(y ~ ffvd(X_se, nbasis = c(10, 10, 10)), data = data)

  # plot the estimated coefficient and its confidence intervals
  # for a selection of curves
  plot_ci(res, beta_index = 1, curves = c(50, 70, 100))
}
}

}
\seealso{
\code{\link{vd_fit}}
}
