% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{Rten2}
\alias{Rten2}
\title{This Rten2 function, is, atm, a mystery to me.}
\usage{
Rten2(X1, X2)
}
\arguments{
\item{X1}{.}

\item{X2}{.}
}
\description{
Auxiliary function found in some functions of the \href{https://cran.r-project.org/web/packages/sommer/index.html}{sommer} package.
All credits to them.
}
\references{
All credits to the \href{https://cran.r-project.org/web/packages/sommer/index.html}{sommer} package authors.
}
