% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simpson.R
\name{Simpson}
\alias{Simpson}
\title{Simpson numeric integration}
\usage{
Simpson(
  fdobj1,
  fdobj2 = NULL,
  fdobj3 = NULL,
  Lfdobj1 = fda::int2Lfd(0),
  Lfdobj2 = fda::int2Lfd(0),
  rng,
  sub = 25,
  wtfd = 0
)
}
\arguments{
\item{fdobj1}{First functional object.}

\item{fdobj2}{Second functional object.}

\item{fdobj3}{Third functional object.}

\item{Lfdobj1}{.}

\item{Lfdobj2}{.}

\item{rng}{.}

\item{sub}{Number of intervals for the Simpson numerical integration method
divided by 2.}

\item{wtfd}{.}
}
\value{
.
}
\description{
Simpson numeric integration proposed in \href{http://dx.doi.org/10.9734/BJMCS/2016/23048}{Burden and Fairies (2016)}
}
