% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VDPO-package.R
\docType{package}
\name{VDPO-package}
\alias{VDPO}
\alias{VDPO-package}
\title{VDPO: Working with and Analyzing Functional Data of Varying Lengths}
\description{
Comprehensive set of tools for analyzing and manipulating functional data with non-uniform lengths. This package addresses two common scenarios in functional data analysis: Variable Domain Data, where the observation domain differs across samples, and Partially Observed Data, where observations are incomplete over the domain of interest. 'VDPO' enhances the flexibility and applicability of functional data analysis in 'R'. See Amaro et al. (2024) \doi{10.48550/arXiv.2401.05839}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pavel-hernadez-amaro.github.io/VDPO/}
}

}
\author{
\strong{Maintainer}: Pavel Hernandez \email{pavel.amaro96@gmail.com}

Other contributors:
\itemize{
  \item Jose Ignacio Diez \email{jodiezru@pa.uc3m.es} [contractor]
  \item Maria Durban \email{mdurban@est-econ.uc3m.es} [contributor]
  \item Maria del Carmen Aguilera-Morillo \email{mdagumor@eio.upv.es} [contributor]
}

}
\keyword{internal}
