% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VDPO.R
\name{VDPO}
\alias{VDPO}
\title{VDPO function}
\usage{
VDPO(formula, data, family = stats::gaussian(), offset = NULL)
}
\arguments{
\item{formula}{.}

\item{data}{.}

\item{family}{.}

\item{offset}{.}
}
\value{
.
}
\description{
VDPO function
}
