% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{VDPO_example_po}
\alias{VDPO_example_po}
\title{Example of partially observed data}
\format{
\subsection{\code{VDPO_example_po}}{

A data frame with 100 rows and 3 columns:
\describe{
\item{y}{}
\item{x}{}
\item{grid}{}
}
}
}
\usage{
VDPO_example_po
}
\description{
This \code{data.frame} shows the structure that a dataset for a partially observed
problem might have
}
\keyword{datasets}
