% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{VDPO_example_vd}
\alias{VDPO_example_vd}
\title{Example of variable domain data}
\format{
\subsection{\code{VDPO_example_vd}}{

A data frame with 100 rows and 7 columns:
\describe{
\item{y}{}
\item{X_s, Y_s}{}
\item{X_se, Y_se}{}
\item{x1}{}
\item{Beta}{}
}
}
}
\usage{
VDPO_example_vd
}
\description{
This \code{data.frame} shows the structure that a dataset for a variable domain
problem might have
}
\keyword{datasets}
