% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XZG2theta.R
\name{XZG2theta}
\alias{XZG2theta}
\title{XZG2theta function}
\usage{
XZG2theta(X, Z, G, TMatrix, y, family = stats::gaussian(), offset = NULL)
}
\arguments{
\item{X}{Fixed part of the mixed model.}

\item{Z}{Random part of the mixed model.}

\item{G}{Variance-covariance matrix.}

\item{y}{Response variable.}

\item{family}{Family of the distrbution.}

\item{offset}{.}

\item{T}{Matrix of transformation from the multivariate model to the mixed model.}
}
\value{
.
}
\description{
XZG2theta function
}
