% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XZG2theta_1d.R
\name{XZG2theta_1d}
\alias{XZG2theta_1d}
\title{One dimensional version of XZG2theta.}
\usage{
XZG2theta_1d(X, Z, G, T, y, family = stats::gaussian())
}
\arguments{
\item{X}{Fixed part of the mixed model.}

\item{Z}{Random part of the mixed model.}

\item{G}{Variance-covariance matrix.}

\item{T}{Matrix of transformation from the multivariate model to the mixed model.}

\item{y}{Response variable.}

\item{family}{Family of the distrbution.}
}
\value{
.
}
\description{
One dimensional version of XZG2theta.
}
