% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data_generator_vd}
\alias{data_generator_vd}
\title{Data generator function for the variable domain case.}
\usage{
data_generator_vd(
  N = 100,
  J = 100,
  nsims = 1,
  Rsq = 0.95,
  aligned = TRUE,
  multivariate = FALSE,
  beta_index = 1,
  use_x = FALSE,
  use_f = FALSE,
  seed = 42
)
}
\arguments{
\item{N}{Number of subjects.}

\item{J}{Number of maximum observations per subject.}

\item{nsims}{Number of simulations per the simulation study.}

\item{Rsq}{Variance of the model.}

\item{aligned}{If the data that will be generated is aligned or not.}

\item{multivariate}{If TRUE, the data is generated with 2 variables.}

\item{beta_index}{Index for the beta.}

\item{use_x}{If the data is generated with x.}

\item{use_f}{If the data is generated with f.}

\item{seed}{Seed for reproducibility.}
}
\value{
Example data.
}
\description{
This function is for internal use.
}
