% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffpo_2d.R
\name{ffpo_2d}
\alias{ffpo_2d}
\title{ffpo_2d}
\usage{
ffpo_2d(
  X_miss,
  miss_points,
  missing_points,
  nbasis = rep(15, 4),
  bdeg = rep(3, 4)
)
}
\arguments{
\item{X_miss}{.}

\item{miss_points}{.}

\item{missing_points}{.}

\item{nbasis}{.}

\item{bdeg}{.}
}
\value{
.
}
\description{
fully functional partially observed for the two dimensional case
}
\details{
The difference between
miss_points and missing_points is the format in which the data is
presented.
}
