% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffvd.R
\name{ffvd}
\alias{ffvd}
\title{Fully functional variable domain function.}
\usage{
ffvd(X, nbasis = c(20, 21, 22), bdeg = c(3, 3, 3))
}
\arguments{
\item{X}{Data matrix.}

\item{nbasis}{.}

\item{bdeg}{.}
}
\value{
.
}
\description{
Fully functional variable domain function.
}
\seealso{
\code{\link{VDPO}}
}
