% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/po_2d_fit.R
\name{po_2d_fit}
\alias{po_2d_fit}
\title{Title}
\usage{
po_2d_fit(formula, data, family = stats::gaussian(), offset = NULL)
}
\arguments{
\item{formula}{.}

\item{data}{.}

\item{family}{.}

\item{offset}{.}
}
\value{
.
}
\description{
Title
}
