% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vd_fit.R
\name{vd_fit}
\alias{vd_fit}
\title{Estimation of the generalized additive functional regression models for
variable domain functional data}
\usage{
vd_fit(formula, data, family = stats::gaussian(), offset = NULL)
}
\arguments{
\item{formula}{a formula object with at least one \code{ffvd} term.}

\item{data}{a \code{list} object containing the response variable
and the covariates.}

\item{family}{a \code{family} object specifying the distribution from which the
data originates. The default distribution is \code{\link{gaussian}}.}

\item{offset}{An offset vector. The default value is \code{NULL}.}
}
\value{
An object of class \code{vd_fit}. It is a \code{list} containing the following items:
\itemize{
\item An item named \code{fit} of class \code{sop}. See \link[SOP]{sop.fit}.
\item An item named \code{Beta} which is the estimated functional coefficient.
\item An item named \code{theta} which is the basis coefficient of \code{Beta}.
\item An item named \code{covar_theta} which is the covariance matrix of \code{theta}.
\item An item named \code{M} which is the number of observations points for each curve.
\item An item named \code{ffvd_evals} which is the result of the evaluations of the \code{ffvd}
terms in the formula.
}
}
\description{
The \code{vd_fit} function fits generalized additive functional regression models
for variable domain functional data.
}
\examples{
\dontrun{
# VARIABLE DOMAIN FUNCTIONAL DATA EXAMPLE

# load the example data
data <- VDPO::VDPO_example_vd

# define a formula object that determines the model behavior
# note that this formula only uses one 'ffvd' term and that
# the 'nbasis' parameter is not the default one
formula <- y ~ ffvd(X_se, nbasis = c(10, 10, 10))

# fit the model with the data and the formula
res <- VDPO(formula = formula, data = data)

# important parameters of the model can be accessed as follows
res$Beta_ffvd         # variable domain functional coefficient
res$fit$fitted.values # estimated response variable

# ------------------------------------------------------------------
# PARTIALLY OBSERVED FUNCTIONAL DATA EXAMPLE

# load the example data
data <- VDPO::VDPO_example_po

# define a formula object that determines the model behavior
# note that this formula only uses one 'ffpo' term
formula <- y ~ ffvd(X_se)

# fit the model with the data and the formula
res <- VDPO(formula = formula, data = data)

# important parameters of the model can be accessed as follows
res$theta_ffpo        # functional coefficient
res$fit$fitted.values # estimated response variable
}
}
\seealso{
\code{\link{ffvd}}, \code{\link{ffpo}}, \code{\link{ffpo_2d}}, \code{\link{add_grid}}
}
