% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VDPO.R
\name{VDPO}
\alias{VDPO}
\title{Estimation of the generalized additive functional regression models for
variable domain and/or partially observed functional regression data.}
\usage{
VDPO(formula, data, family = stats::gaussian(), offset = NULL)
}
\arguments{
\item{formula}{a formula object with at least one \code{ffvd}, \code{ffpo} or
\code{ffpo_2d} term.}

\item{data}{a \code{data.frame} object containing the response variable
and the covariates. When fitting partially observed functional data,
a grid of observation points is also needed.}

\item{family}{a \code{family} object specifying the distribution from which the
data originates. The default distribution is \code{\link{gaussian}}.}

\item{offset}{An offset vector. The default value is \code{NULL}.}
}
\value{
Object of class \code{VDPO} with the results of the computation.
}
\description{
The \code{VDPO} function fits generalized additive functional regression models
for variable domain and partially observed functional data in both 1 and 2 dimensions.
}
\seealso{
\code{\link{ffvd}}, \code{\link{ffpo}}, \code{\link{ffpo_2d}}, \code{\link{addgrid}}
}
