% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VDPO.R
\name{VDPO}
\alias{VDPO}
\title{Variable Domain Partially observed function.}
\usage{
VDPO(formula, data, family = stats::gaussian(), offset = NULL)
}
\arguments{
\item{formula}{A formula object.}

\item{data}{A data-frame object.}

\item{family}{A family object specifying the distribution from which the
data originates. The default distribution is \code{stats::gaussian}.}

\item{offset}{An offset vector. The default value is \code{NULL}.}
}
\value{
Object of class \code{VDPO} with the results of the computation.
}
\description{
This function does ......
}
